Hardcoded images to Aztecs and later will do the rest, also added an imagePath to the Civilization class
Added a AssignedPicture function to create a switch statement with a return value, created a iteration for a variable called NewCivList from ObservableCollection Civilization List, and created a foreach item in civilizations called the imagepath and assigned the value of AssignedPicture with the civ.name parameter which are properties of type string that were in the list of name's string. We then call the NewCivList and add the civ parameter to it and after the foreach we equal the Civilization List with the NewCivList. Inside the AssignedPicture we add a string civName parameter to then use it in the switch statement, created a variable called CivImage with an empty string. Each case has the name of a Civilization to assign the image that we want to hardcode to that civilization and we return the CivImage value to return it to the civ.imagePath and give "Aztecs" it's image within the label display of each civilization.